backend/pyproject.toml
[project]
name = "blockbattery-backend"
version = "0.1.0"
description = "BlockBattery backend: submission verification, mock receipt tokens, and downstream modeling APIs"
requires-python = ">=3.12"
dependencies = [
    "entsoe-py>=0.6.16",
    "httpx>=0.27",
    "pandas>=2.2",
    "pyarrow>=16",
    "cvxpy>=1.5",
    "highspy>=1.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-dotenv>=1.0",
    "numpy>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
